Allow -l <float>
The option -l takes the sequence length as its argument. So far, this
had to be an integer. Now it can be a floating point number, allowing
the convenient exponent notation as in -l 3e9 to specify the length of
the human genome as 3 billion nucleotides.

+++ srcEpos/interface.c
@@ -66,7 +66,7 @@ Args *getArgs(int argc, char *argv[]){
     switch(c){
       break;
     case 'l':                           /* sequence length */
-      args->l = atol(optarg);
+      args->l = (long)atof(optarg);
       break;
     case 'L':                           /* preset levels */
       args->L = estrdup(optarg);
